Multiple sync'd view models w/ immutable model
Because the models we are working with contain a lot computed
values that need to be rendered into view, and changes to one of
the dependent values in one view needs to be updated in another,
there needs to be a method to flow those change notifications.

The simplest way is to make the model immutable, and then
trigger an update when a new model is available with the new values.
This means any update anywhere will create an entirely new
version of the model that is flowed to all other views.

Other alternatives would require dependency tracking between all
the various game variables, which are dynamically generated.
While that would be more efficient for modifications, it means
that the model needs to now have change notifications built in,
which greatly complicates the logic.